adding search impl through tree-sitter-query also add tests
