+++ gen/tree-sitter/query/src/lib.rs
@@ -49,4 +49,4 @@ mod tnode {
 #[cfg(feature = "legion")]
 pub use tnode::TNode;
 
-
+mod search;
